Application feedback and logging changes
Debug levels were merged into a single "-log" or "-verbose" command line parameter to support adding in more levels. Each level adds in a new category:
- 0 - none
- 1 - com.cloudant classes
- 2 - org.lightcouch classes
- 3 - org.apache.http classes

INFO and above messages are logged to stdout as application feedback. All messages are logged to a file now and saved for each run of the program.